omni_proj/requirement.txt
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
channels==1.1.3
constantly==15.1.0
daphne==1.2.0
Django==1.11.13
django-filter==1.1.0
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
Markdown==2.6.11
msgpack-python==0.5.6
pytz==2018.4
redis==2.10.6
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
zope.interface==4.5.0